requirements.txt
langchain
langchain-community
langgraph
python-dotenv
streamlit
openai
faiss-cpu
tiktoken
unstructured
